Fix | Client | Extract views components
