+++ client/src/components/App.js
@@ -1,21 +1,22 @@
 import React from 'react';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
-import Header from './Header';
-import HomePage from './HomePage';
-import SignIn from './SignIn';
-import RegisterUser from './RegisterUser';
-import SingleTrip from './SingleTrip';
-import AddTrip from './AddTrip';
-import EditTrip from './EditTrip';
-import EditExpense from './EditExpense';
-import AddExpense from './AddExpense';
-import TripSummary from './TripSummary';
-import AllTrips from './AllTrips';
-import UserProfile from './UserProfile';
-import CurrenciesRates from './CurrenciesRates';
-import AllExpenses from './AllExpenses';
-import Footer from './Footer';
+
 import Layout from '../layout/Layout';
+import CurrenciesRates from '../views/CurrenciesRates/CurrenciesRates.component';
+import ExpenseAdd from '../views/ExpenseAdd/ExpenseAdd.component';
+import ExpenseEdit from '../views/ExpenseEdit/ExpenseEdit.component';
+import ExpensesAll from '../views/ExpensesAll/ExpensesAll.component';
+import HomePage from '../views/HomePage/HomePage.component';
+import RegisterUser from '../views/RegisterUser/RegisterUser.component';
+import SignIn from '../views/SignIn/SignIn.component';
+import TripAdd from '../views/TripAdd/TripAdd.component';
+import TripEdit from '../views/TripEdit/TripEdit.component';
+import TripsAll from '../views/TripsAll/TripsAll.component';
+import TripSingle from '../views/TripSingle/TripSingle.component';
+import TripSummary from '../views/TripSummary/TripSummary.component';
+import UserProfile from '../views/UserProfile/UserProfile.component';
+import Footer from './Footer';
+import Header from './Header';
 import PrivateRoute from './PrivateRoute';
 
 class App extends React.Component {
@@ -25,16 +26,16 @@ class App extends React.Component {
         <Layout>
           <Header />
           <Route path="/" exact component={HomePage} />
-          <PrivateRoute path="/trips/add" component={AddTrip} />
-          <PrivateRoute path="/trips/edit/:id" component={EditTrip} />
+          <PrivateRoute path="/trips/add" component={TripAdd} />
+          <PrivateRoute path="/trips/edit/:id" component={TripEdit} />
           <Route path="/users/login" component={SignIn} />
           <Route path="/users/register" component={RegisterUser} />
-          <PrivateRoute path="/trips/single/:tripId" component={SingleTrip} />
+          <PrivateRoute path="/trips/single/:tripId" component={TripSingle} />
           <PrivateRoute path="/trips/summary/:id" component={TripSummary} />
-          <PrivateRoute path="/trips/all" component={AllTrips} />
-          <PrivateRoute path="/trips/:tripId/expenses/edit/:expenseId" component={EditExpense} />
-          <PrivateRoute path="/trips/:tripId/expenses/add" component={AddExpense} />
-          <PrivateRoute path="/trips/:tripId/expenses/all" component={AllExpenses} />
+          <PrivateRoute path="/trips/all" component={TripsAll} />
+          <PrivateRoute path="/trips/:tripId/expenses/edit/:expenseId" component={ExpenseEdit} />
+          <PrivateRoute path="/trips/:tripId/expenses/add" component={ExpenseAdd} />
+          <PrivateRoute path="/trips/:tripId/expenses/all" component={ExpensesAll} />
           <PrivateRoute path="/users/profile" component={UserProfile} />
           <PrivateRoute path="/trips/currencies/:tripId" component={CurrenciesRates} />
           <Footer />
