--- client/src/components/RatesList.js
@@ -1,277 +1,0 @@
-export const ratesObject = {
-  date: '2020-01-07',
-  data: {
-    PLN: [
-      {
-        name: 'PLN',
-        rate: 1,
-      },
-      {
-        name: 'USD',
-        rate: 3.799447,
-      },
-      {
-        name: 'GBP',
-        rate: 4.976906,
-      },
-      {
-        name: 'EUR',
-        rate: 4.245571,
-      },
-      {
-        name: 'JPY',
-        rate: 0.035165,
-      },
-      {
-        name: 'AUD',
-        rate: 2.640084,
-      },
-      {
-        name: 'CAD',
-        rate: 2.926855,
-      },
-      {
-        name: 'CHF',
-        rate: 3.914324,
-      },
-    ],
-    USD: [
-      {
-        name: 'PLN',
-        rate: 0.263199,
-      },
-      {
-        name: 'USD',
-        rate: 1,
-      },
-      {
-        name: 'GBP',
-        rate: 1.311947,
-      },
-      {
-        name: 'EUR',
-        rate: 1.11767,
-      },
-      {
-        name: 'JPY',
-        rate: 0.009248,
-      },
-      {
-        name: 'AUD',
-        rate: 0.694455,
-      },
-      {
-        name: 'CAD',
-        rate: 0.770602,
-      },
-      {
-        name: 'CHF',
-        rate: 1.030908,
-      },
-    ],
-    GBP: [
-      {
-        name: 'PLN',
-        rate: 0.200929,
-      },
-      {
-        name: 'USD',
-        rate: 0.762233,
-      },
-      {
-        name: 'GBP',
-        rate: 1,
-      },
-      {
-        name: 'EUR',
-        rate: 0.852071,
-      },
-      {
-        name: 'JPY',
-        rate: 0.007049,
-      },
-      {
-        name: 'AUD',
-        rate: 0.529338,
-      },
-      {
-        name: 'CAD',
-        rate: 0.587377,
-      },
-      {
-        name: 'CHF',
-        rate: 0.785788,
-      },
-    ],
-    EUR: [
-      {
-        name: 'PLN',
-        rate: 0.23554,
-      },
-      {
-        name: 'USD',
-        rate: 0.894721,
-      },
-      {
-        name: 'GBP',
-        rate: 1.173614,
-      },
-      {
-        name: 'EUR',
-        rate: 1,
-      },
-      {
-        name: 'JPY',
-        rate: 0.008274,
-      },
-      {
-        name: 'AUD',
-        rate: 0.621343,
-      },
-      {
-        name: 'CAD',
-        rate: 0.689465,
-      },
-      {
-        name: 'CHF',
-        rate: 0.922274,
-      },
-    ],
-    JPY: [
-      {
-        name: 'PLN',
-        rate: 28.437415,
-      },
-      {
-        name: 'USD',
-        rate: 108.133995,
-      },
-      {
-        name: 'GBP',
-        rate: 141.866535,
-      },
-      {
-        name: 'EUR',
-        rate: 120.861798,
-      },
-      {
-        name: 'JPY',
-        rate: 1,
-      },
-      {
-        name: 'AUD',
-        rate: 75.094096,
-      },
-      {
-        name: 'CAD',
-        rate: 83.328356,
-      },
-      {
-        name: 'CHF',
-        rate: 111.476309,
-      },
-    ],
-    AUD: [
-      {
-        name: 'PLN',
-        rate: 0.378779,
-      },
-      {
-        name: 'USD',
-        rate: 1.43998,
-      },
-      {
-        name: 'GBP',
-        rate: 1.889182,
-      },
-      {
-        name: 'EUR',
-        rate: 1.609426,
-      },
-      {
-        name: 'JPY',
-        rate: 0.013317,
-      },
-      {
-        name: 'AUD',
-        rate: 1,
-      },
-      {
-        name: 'CAD',
-        rate: 1.109653,
-      },
-      {
-        name: 'CHF',
-        rate: 1.484487,
-      },
-    ],
-    CAD: [
-      {
-        name: 'PLN',
-        rate: 0.341667,
-      },
-      {
-        name: 'USD',
-        rate: 1.297687,
-      },
-      {
-        name: 'GBP',
-        rate: 1.702493,
-      },
-      {
-        name: 'EUR',
-        rate: 1.450403,
-      },
-      {
-        name: 'JPY',
-        rate: 0.012001,
-      },
-      {
-        name: 'AUD',
-        rate: 0.901186,
-      },
-      {
-        name: 'CAD',
-        rate: 1,
-      },
-      {
-        name: 'CHF',
-        rate: 1.337794,
-      },
-    ],
-    CHF: [
-      {
-        name: 'PLN',
-        rate: 0.255473,
-      },
-      {
-        name: 'USD',
-        rate: 0.970022,
-      },
-      {
-        name: 'GBP',
-        rate: 1.272611,
-      },
-      {
-        name: 'EUR',
-        rate: 1.084277,
-      },
-      {
-        name: 'JPY',
-        rate: 0.008971,
-      },
-      {
-        name: 'AUD',
-        rate: 0.673637,
-      },
-      {
-        name: 'CAD',
-        rate: 0.7475,
-      },
-      {
-        name: 'CHF',
-        rate: 1,
-      },
-    ],
-  },
-};
