+++ client/src/views/ExpenseAdd/ExpenseAdd.component.jsx
@@ -1,24 +1,25 @@
-import React, { Component } from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import axios from 'axios';
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import ContentWrapper from './ContentWrapper';
-import Button from './Button';
+import Select from 'react-select';
+
+import Button from '../../components/Button';
+import ContentWrapper from '../../components/ContentWrapper';
+import ErrorMessage from '../../components/ErrorMessage';
 import {
-  Form,
-  Label,
-  Input,
   customStyleSelect,
-  TripHeader,
+  Form,
+  Input,
+  Label,
   LinkText,
   NavLinksContainer,
-} from './styled';
-import Select from 'react-select';
-import getToken from '../utils/getToken';
-import formatCurrencies from '../utils/formatCurrencies';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import ErrorMessage from './ErrorMessage';
+  TripHeader,
+} from '../../components/styled';
+import formatCurrencies from '../../utils/formatCurrencies';
+import getToken from '../../utils/getToken';
 
-class AddExpense extends Component {
+class ExpenseAdd extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -197,4 +198,4 @@ const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps)(AddExpense);
+export default connect(mapStateToProps)(ExpenseAdd);
