+++ client/src/views/ExpenseEdit/ExpenseEdit.component.jsx
@@ -1,23 +1,24 @@
-import React, { Component } from 'react';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import axios from 'axios';
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import ContentWrapper from './ContentWrapper';
-import Button from './Button';
+import Select from 'react-select';
+
+import Button from '../../components/Button';
+import ContentWrapper from '../../components/ContentWrapper';
 import {
-  Form,
-  Label,
-  Input,
   customStyleSelect,
-  TripHeader,
+  Form,
+  Input,
+  Label,
   LinkText,
   NavLinksContainer,
-} from './styled';
-import Select from 'react-select';
-import getToken from '../utils/getToken';
-import formatCurrencies from '../utils/formatCurrencies';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+  TripHeader,
+} from '../../components/styled';
+import formatCurrencies from '../../utils/formatCurrencies';
+import getToken from '../../utils/getToken';
 
-class EditExpense extends Component {
+class ExpenseEdit extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -234,4 +235,4 @@ const mapStateToProps = (state) => {
   };
 };
 
-export default connect(mapStateToProps)(EditExpense);
+export default connect(mapStateToProps)(ExpenseEdit);
